+++ gpgutil/cipherfile.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 	"io"
 
-	"golang.org/x/crypto/openpgp"
-	"golang.org/x/crypto/openpgp/packet"
+	"github.com/ProtonMail/go-crypto/openpgp"
+	"github.com/ProtonMail/go-crypto/openpgp/packet"
 )
 
 // Encrypt encrypts stream into an OpenPGP-encrypted stream